src/main/java/vip/iotworld/service/Impl/CategoryServiceImpl.java
package vip.iotworld.service.Impl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import vip.iotworld.dataobject.ProductCategory;
import vip.iotworld.repository.ProductCategoryRepository;
import vip.iotworld.service.CategoryService;

import java.util.List;

/**
 * Created with IDEA
 * author:LiuBing
 * Date:17/08/2018
 * Time:15:02
 */
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> finaAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
